split board to board and board_random
